finalisation conversion map()
