+++ dbmongo/js/reduce.algo2/dealWithProcols.ts
@@ -8,7 +8,7 @@ type OutputEvent = {
   date_proc_col: Date
 }
 
-export type Output = {
+type Output = {
   etat_proc_collective: unknown
   date_proc_collective: unknown
   tag_failure: boolean
@@ -28,9 +28,9 @@ export function dealWithProcols(
       const the_event = data_source[hash]
 
       let etat: AltaresToHumanRes | ProcolToHumanRes = null
-      if (altar_or_procol == "altares")
+      if (altar_or_procol === "altares")
         etat = f.altaresToHuman(the_event.code_evenement)
-      else if (altar_or_procol == "procol")
+      else if (altar_or_procol === "procol")
         etat = f.procolToHuman(the_event.action_procol, the_event.stade_procol)
 
       if (etat != null)
@@ -65,7 +65,7 @@ export function dealWithProcols(
       if (time in output_indexed) {
         output_indexed[time].etat_proc_collective = event.etat
         output_indexed[time].date_proc_collective = event.date_proc_col
-        if (event.etat != "in_bonis") output_indexed[time].tag_failure = true
+        if (event.etat !== "in_bonis") output_indexed[time].tag_failure = true
       }
     })
   })
